Refactor to Golang
This effort rewrites all micro-services into a single Golang monolith.
I hope to improve stability by cutting all REST connections.
Next to this, it cuts the use of Docker in favor of compiling the binary directly on the end device for best compatibility.
Old micro-services setup is still available in branch "deprecated/0-micro-services".